Fix deposit to Perps market and separate test steps for easier copypaste
